Race PRs: show pace, even card sizing; fix cos/sin ambiguity
- Race PR cards + history now show average pace (min/mi), derived from
  time and distance (no backend change)
- MADStatCard gains an optional subtitle; empty PR cards reserve the pace
  line so all cards are equal height
- Fix "Ambiguous use of 'cos'" in FeedMediaViews (wrap cos/sin results in
  CGFloat; newer toolchains don't bridge Double<->CGFloat implicitly)

+++ app/Mile A Day/Views/Feed/FeedMediaViews.swift
@@ -438,7 +438,7 @@ struct HypeBurstView: View {
 
     private func particleOffset(angle degrees: Double, distance: CGFloat) -> CGSize {
         let radians = degrees * .pi / 180
-        return CGSize(width: cos(radians) * distance, height: sin(radians) * distance)
+        return CGSize(width: CGFloat(cos(radians)) * distance, height: CGFloat(sin(radians)) * distance)
     }
 
     private func play() {
